requirements.txt
python-telegram-bot~=20.0
environs~=9.5
